+++ e2e/__tests__/fakePromises.test.ts
@@ -8,13 +8,8 @@
 import runJest from '../runJest';
 
 describe('Fake promises', () => {
-  it('should be possible to resolve with fake timers using immediates', () => {
-    const result = runJest('fake-promises/immediate');
-    expect(result.status).toBe(0);
-  });
-
-  it('should be possible to resolve with fake timers using asap', () => {
-    const result = runJest('fake-promises/asap');
+  it('should be possible to resolve with fake timers', () => {
+    const result = runJest('fake-promises');
     expect(result.status).toBe(0);
   });
 });